requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
email-validator==2.1.0
pydantic[email]==2.5.0
psycopg2-binary==2.9.9
python-slugify==8.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
# Image processing dependencies
Pillow==10.1.0
aiofiles==23.2.1
# For debugging and linting
mypy==1.5.1 